+++ History.md
@@ -1,3 +1,10 @@
+1.2.0 / 2020-02-12
+==================
+
+  * Adapted function `getEnumerationMembers()` to the Enumeration API (now has 'implicit' as parameter).
+  * Adapted function `getDomainEnumerations()` to the EnumerationMember API (now has 'implicit' as parameter).
+  * Adapted function `toJSON()` to the EnumerationMember API (now has 'implicit' as parameter).
+
 1.1.0 / 2020-02-04
 ==================